+++ onmt/utils/Parallel.lua
@@ -4,7 +4,6 @@
 ]]--
 
 local Parallel = {
-  gpus = {0},
   _pool = nil,
   count = 1,
   gradBuffer = torch.Tensor()
@@ -31,25 +30,26 @@ end
 
 function Parallel.init(opt)
   if onmt.utils.Cuda.activated then
-    Parallel.count = opt.nparallel
-    Parallel.gpus = onmt.utils.Cuda.getGPUs(opt.nparallel)
+    Parallel.count = onmt.utils.Cuda.gpuCount()
     Parallel.gradBuffer = onmt.utils.Cuda.convert(Parallel.gradBuffer)
     Parallel._tds = require('tds')
 
     if Parallel.count > 1 then
-      print('Using ' .. Parallel.count .. ' threads on ' .. #Parallel.gpus .. ' GPUs')
+      local globalLogger = _G.logger
       local threads = require('threads')
       threads.Threads.serialization('threads.sharedserialize')
-      local thegpus = Parallel.gpus
       Parallel._gmutex = threads.Mutex()
       Parallel._pool = threads.Threads(
         Parallel.count,
-        function(threadid)
+        function()
           require('cunn')
           require('nngraph')
           require('onmt.init')
           _G.threads = require('threads')
-          onmt.utils.Cuda.init(opt, thegpus[threadid])
+        end,
+        function(threadid)
+          _G.logger = globalLogger
+          onmt.utils.Cuda.init(opt, threadid)
         end
       ) -- dedicate threads to GPUs
       Parallel._pool:specific(true)
@@ -60,22 +60,15 @@ function Parallel.init(opt)
       local ret
       ret, Parallel.usenccl = pcall(require, 'nccl')
       if not ret then
-        print("WARNING: for improved efficiency in nparallel mode - do install nccl")
+        _G.logger:warning("For improved efficiency with multiple GPUs, consider installing nccl")
         Parallel.usenccl = nil
       elseif os.getenv('CUDA_LAUNCH_BLOCKING') == '1' then
-        print("WARNING: CUDA_LAUNCH_BLOCKING set - cannot use nccl")
+        _G.logger:warning("CUDA_LAUNCH_BLOCKING set - cannot use nccl")
         Parallel.usenccl = nil
       end
     end
 
   end
-end
-
-function Parallel.getGPU(i)
-  if onmt.utils.Cuda.activated and Parallel.gpus[i] ~= 0 then
-    return Parallel.gpus[i]
-  end
-  return 0
 end
 
 --[[ Launch function in parallel on different threads. ]]
@@ -105,10 +98,10 @@ function Parallel.accGradParams(gradParams, batches)
 
          -- Synchronize before and after copy to ensure that it doesn't overlap
          -- with this add or previous adds
-          waitForDevice(Parallel.gpus[j], Parallel.gpus[1])
+          waitForDevice(onmt.utils.Cuda.gpuIds[j], onmt.utils.Cuda.gpuIds[1])
           local remoteGrads = onmt.utils.Tensor.reuseTensor(Parallel.gradBuffer, gradParams[j][h]:size())
           remoteGrads:copy(gradParams[j][h])
-          waitForDevice(Parallel.gpus[1], Parallel.gpus[j])
+          waitForDevice(onmt.utils.Cuda.gpuIds[1], onmt.utils.Cuda.gpuIds[j])
           gradParams[1][h]:add(remoteGrads)
         else
           table.insert(inputs, gradParams[j][h])
@@ -151,7 +144,7 @@ function Parallel.syncParams(params)
         for h = 1, #params[1] do
           params[j][h]:copy(params[1][h])
         end
-        waitForDevice(Parallel.gpus[j], Parallel.gpus[1])
+        waitForDevice(onmt.utils.Cuda.gpuIds[j], onmt.utils.Cuda.gpuIds[1])
       end
     else
       for h = 1, #params[1] do